pyproject.toml
[tool.poetry]
name = "lesson27_solution"
version = "0.1.0"
description = ""
authors = ["Olga <olgabryankina5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.2"
Django = "^4.1.2"
psycopg2 = "^2.9.5"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
